+++ lib/tests/processing.rs
@@ -32,12 +32,12 @@ fn arg_list_owned_set() {
     // Test works (compiles) using a `String` based slice (as given from `evn::args()` for real args)
     // Note, **deliberately** not using the `arg_list` macro here!
     let args: Vec<String> = vec![ String::from("--foo"), String::from("--bah") ];
-    let _ = gong::process(&args, &get_base());
+    let _ = get_base().process(&args);
 
     // Test works (compiles) using a `&str` based slice
     // Note, **deliberately** not using the `arg_list` macro here!
     let args: Vec<&str> = vec![ "--foo", "--bah" ];
-    let _ = gong::process(&args, &get_base());
+    let _ = get_base().process(&args);
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -88,7 +88,7 @@ fn basic() {
             expected_item!(13, NonOption, "jkl"),
         ]
     );
-    check_result(&Actual(gong::process(&args, &get_base())), &expected);
+    check_result(&Actual(get_base().process(&args)), &expected);
 }
 
 /// Test that everything after an early terminator is taken to be a non-option, including any
@@ -125,7 +125,7 @@ fn early_term() {
             expected_item!(14, NonOption, "-b"),
         ]
     );
-    check_result(&Actual(gong::process(&args, &get_base())), &expected);
+    check_result(&Actual(get_base().process(&args)), &expected);
 }
 
 /// Test empty long option names with data param (-- on it's own is obviously picked up as early
@@ -141,7 +141,7 @@ fn long_no_name() {
             expected_item!(1, LongWithNoName),
         ]
     );
-    check_result(&Actual(gong::process(&args, &get_base())), &expected);
+    check_result(&Actual(get_base().process(&args)), &expected);
 }
 
 /// Test repetition - each instance should exist in the results in its own right. Note, data arg
@@ -165,7 +165,7 @@ fn repetition() {
             expected_item!(7, Short, 'h'),
         ]
     );
-    check_result(&Actual(gong::process(&args, &get_base())), &expected);
+    check_result(&Actual(get_base().process(&args)), &expected);
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -192,7 +192,7 @@ mod utf8 {
                 expected_item!(4, Short, '❤'), // '\u{2764}' black heart
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Some utf8 multi-byte char handling - chars with combinator chars (e.g. accent)
@@ -218,7 +218,7 @@ mod utf8 {
                 expected_item!(7, Long, "ábc"),              // with combinator
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Some utf8 multi-byte char width handling - chars with variation selector
@@ -239,7 +239,7 @@ mod utf8 {
                 expected_item!(2, UnknownLong, "❤️"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Some utf8 multi-byte char width handling - lone combinator chars
@@ -256,7 +256,7 @@ mod utf8 {
                 expected_item!(3, Short, '\u{030a}'),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 }
 
@@ -282,7 +282,7 @@ mod abbreviations {
                 expected_item!(1, AmbiguousLong, "fo"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test handling of abbreviated long options, without ambiguity
@@ -304,7 +304,7 @@ mod abbreviations {
                 expected_item!(3, Long, "foobar"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test handling when abbreviated matching is disabled
@@ -323,9 +323,9 @@ mod abbreviations {
                 expected_item!(5, Long, "foobar"),
             ]
         );
-        let mut opts = get_base();
+        let mut opts = get_base().clone();
         opts.set_allow_abbreviations(false);
-        check_result(&Actual(gong::process(&args, &opts)), &expected);
+        check_result(&Actual(opts.process(&args)), &expected);
     }
 
     /// Test that an exact match overrides ambiguity
@@ -354,7 +354,7 @@ mod abbreviations {
             ],
             vec![]
         );
-        check_result(&Actual(gong::process(&args, &opts)), &expected);
+        check_result(&Actual(opts.process(&args)), &expected);
     }
 }
 
@@ -384,7 +384,7 @@ mod data {
                 expected_item!(4, Long, "help"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test calculation of whether or not short-opt taking data is the last character in the short
@@ -404,7 +404,7 @@ mod data {
                 expected_item!(1, NonOption, "g"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test option with expected data arg, provided in next argument for short options
@@ -432,7 +432,7 @@ mod data {
                 expected_item!(6, ShortWithData, 'o', "def", DataLocation::NextArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test option with expected data arg, provided in same argument for short options
@@ -475,7 +475,7 @@ mod data {
                 expected_item!(9, ShortWithData, 'o', "abx", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test missing argument data for long option
@@ -489,7 +489,7 @@ mod data {
                 expected_item!(0, LongMissingData, "hah"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test missing argument data for short option
@@ -506,7 +506,7 @@ mod data {
                 expected_item!(0, ShortMissingData, 'o'),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test some misc. data handling.
@@ -545,7 +545,7 @@ mod data {
                 expected_item!(8, ShortWithData, 'o', "=b", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test repetition - each instance should exist in the results in its own right. Note, basic
@@ -563,7 +563,7 @@ mod data {
                 expected_item!(4, ShortWithData, 'o', "bc", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test option with expected data arg, declared to be in same argument, but empty
@@ -585,7 +585,7 @@ mod data {
                 expected_item!(3, NonOption, "help"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test option with expected data arg, with data containing '='. An '=' in a long option arg
@@ -619,7 +619,7 @@ mod data {
                 expected_item!(10, ShortWithData, 'o', "===o", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test argument data that looks like options
@@ -658,7 +658,7 @@ mod data {
                 expected_item!(22, ShortWithData, 'o', "--blah", DataLocation::NextArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test argument data that looks like early terminator
@@ -680,7 +680,7 @@ mod data {
                 expected_item!(5, ShortWithData, 'o', "--", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test long option involving multi-byte chars, to ensure "in-arg" component splitting for
@@ -699,7 +699,7 @@ mod data {
                 expected_item!(5, LongMissingData, "ƒƒ"),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test short options involving multi-byte chars to check offset calculations in iterating
@@ -741,7 +741,7 @@ mod data {
                 expected_item!(13, ShortWithData, 'Ɛ', "b❤", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 
     /// Test the effect of Utf-8 combinator characters - does this break char iteration or byte
@@ -786,7 +786,7 @@ mod data {
                 expected_item!(6, ShortWithData, 'Ɛ', "\u{030a}b❤\u{fe0f}", DataLocation::SameArg),
             ]
         );
-        check_result(&Actual(gong::process(&args, &get_base())), &expected);
+        check_result(&Actual(get_base().process(&args)), &expected);
     }
 }
 
@@ -846,7 +846,7 @@ mod alt_mode {
         );
         let mut opts = get_base();
         opts.set_mode(OptionsMode::Alternate);
-        check_result(&Actual(gong::process(&args, &opts)), &expected);
+        check_result(&Actual(opts.process(&args)), &expected);
     }
 
     /// Check unexpected and missing data
@@ -868,7 +868,7 @@ mod alt_mode {
         );
         let mut opts = get_base();
         opts.set_mode(OptionsMode::Alternate);
-        check_result(&Actual(gong::process(&args, &opts)), &expected);
+        check_result(&Actual(opts.process(&args)), &expected);
     }
 
     /// Test argument data that looks like early terminator
@@ -888,6 +888,6 @@ mod alt_mode {
         );
         let mut opts = get_base();
         opts.set_mode(OptionsMode::Alternate);
-        check_result(&Actual(gong::process(&args, &opts)), &expected);
+        check_result(&Actual(opts.process(&args)), &expected);
     }
 }