+++ lib/CHANGELOG.md
@@ -2,7 +2,7 @@
 
  * Refactored and optimised the analysis engine
  * Added a `process` method to `Options` as a cleaner alternative to calling the `process` function
-   directly.
+   directly. Deprecated direct use of the `process` function.
 
 # 1.1.2 (October 26th, 2018)